# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for "Мечты добрых сердец" charity foundation
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: public
  - name: forms
  - name: admin
  - name: settings
  - name: Storage
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadUrlRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }

  /visits:
    post:
      operationId: trackVisit
      tags: [forms]
      summary: Record a page visit (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }

  /settings:
    get:
      operationId: getPublicSettings
      tags: [public]
      summary: Public site settings (org info, payment links, contacts)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicSettings" }

  /children:
    get:
      operationId: listChildren
      tags: [public]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Child" }

  /children/{id}:
    get:
      operationId: getChild
      tags: [public]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
        "404":
          description: Not found

  /stories:
    get:
      operationId: listStories
      tags: [public]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Story" }

  /reports:
    get:
      operationId: listReports
      tags: [public]
      summary: Reports for the current week (PDFs published in last 7 days)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }

  /reports/archive:
    get:
      operationId: listArchiveReports
      tags: [public]
      summary: Archived reports older than 7 days
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }

  /donation-clicks:
    post:
      operationId: trackDonationClick
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childId]
              properties:
                childId: { type: integer }
                amount: { type: integer, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /volunteers:
    post:
      operationId: submitVolunteer
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VolunteerInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /materials:
    post:
      operationId: submitMaterialHelp
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialHelpInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /help-requests:
    post:
      operationId: submitHelpRequest
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HelpRequestInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /contacts:
    post:
      operationId: submitContact
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminMe" }
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminMe" }
        "401":
          description: Unauthorized

  /admin/children:
    get:
      operationId: adminListChildren
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Child" }
    post:
      operationId: adminCreateChild
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChildInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }

  /admin/children/{id}:
    put:
      operationId: adminUpdateChild
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChildInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Child" }
    delete:
      operationId: adminDeleteChild
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /admin/stories:
    get:
      operationId: adminListStories
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Story" }
    post:
      operationId: adminCreateStory
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoryInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Story" }

  /admin/stories/{id}:
    put:
      operationId: adminUpdateStory
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoryInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Story" }
    delete:
      operationId: adminDeleteStory
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Report" }
    post:
      operationId: adminCreateReport
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }

  /admin/reports/{id}:
    delete:
      operationId: adminDeleteReport
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }

  /admin/submissions:
    get:
      operationId: adminListSubmissions
      tags: [admin]
      summary: All form submissions in one feed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSubmissions" }

  /admin/submissions/{type}/{id}/archive:
    post:
      operationId: adminArchiveSubmission
      tags: [admin]
      summary: Move a submission to the archive
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [volunteers, materials, helpRequests, contacts]
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }

  /admin/donation-stats:
    get:
      operationId: adminDonationStats
      tags: [admin]
      summary: Weekly donation click counts grouped by child + recent click list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DonationStats" }

  /admin/visit-stats:
    get:
      operationId: adminVisitStats
      tags: [admin]
      summary: Daily visitor counts and recent visits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitStats" }

  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicSettings" }
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicSettings" }

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        id: { type: integer, nullable: true }

    Child:
      type: object
      required: [id, name, surname, age, diagnosis, story, photoUrl, targetSum, collectedSum, isActive, isUrgent]
      properties:
        id: { type: integer }
        name: { type: string }
        surname: { type: string }
        age: { type: integer }
        diagnosis: { type: string }
        story: { type: string }
        photoUrl: { type: string }
        targetSum: { type: integer }
        collectedSum: { type: integer }
        isActive: { type: boolean }
        isUrgent: { type: boolean }

    ChildInput:
      type: object
      required: [name, surname, age, diagnosis, story, photoUrl, targetSum, collectedSum]
      properties:
        name: { type: string }
        surname: { type: string }
        age: { type: integer }
        diagnosis: { type: string }
        story: { type: string }
        photoUrl: { type: string }
        targetSum: { type: integer }
        collectedSum: { type: integer }
        isActive: { type: boolean }
        isUrgent: { type: boolean }

    Story:
      type: object
      required: [id, title, description, photoUrl]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        photoUrl: { type: string }

    StoryInput:
      type: object
      required: [title, description, photoUrl]
      properties:
        title: { type: string }
        description: { type: string }
        photoUrl: { type: string }

    Report:
      type: object
      required: [id, title, fileUrl, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        fileUrl: { type: string }
        createdAt: { type: string, format: date-time }

    ReportInput:
      type: object
      required: [title, fileUrl]
      properties:
        title: { type: string }
        fileUrl: { type: string }

    VolunteerInput:
      type: object
      required: [name, phone, email, city, helpType]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: string }
        city: { type: string }
        helpType: { type: string }

    MaterialHelpInput:
      type: object
      required: [name, phone, items, preferredDate]
      properties:
        name: { type: string }
        phone: { type: string }
        items: { type: string }
        preferredDate: { type: string }

    HelpRequestInput:
      type: object
      required: [childName, age, diagnosis, targetSum, parentContacts]
      properties:
        childName: { type: string }
        age: { type: integer }
        diagnosis: { type: string }
        targetSum: { type: integer }
        parentContacts: { type: string }
        photoUrl: { type: string, nullable: true }

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string }
        email: { type: string }
        message: { type: string }

    AdminMe:
      type: object
      required: [username]
      properties:
        username: { type: string }

    PublicSettings:
      type: object
      required: [orgName, slogan, mission, legalAddress, email, phone, donationLink, requisites, socials, documents, logoUrl, logoSize, paymentQrUrl, adminEmail]
      properties:
        orgName: { type: string }
        slogan: { type: string }
        mission: { type: string }
        legalAddress: { type: string }
        email: { type: string }
        phone: { type: string }
        adminEmail: { type: string }
        donationLink: { type: string }
        requisites: { type: string }
        logoUrl: { type: string }
        logoSize: { type: integer }
        paymentQrUrl: { type: string }
        socials:
          type: object
          properties:
            telegram: { type: string }
            vk: { type: string }
            whatsapp: { type: string }
            max: { type: string }
            instagram: { type: string }
        documents:
          type: array
          items:
            type: object
            required: [title, url]
            properties:
              title: { type: string }
              url: { type: string }

    SettingsInput:
      $ref: "#/components/schemas/PublicSettings"

    AdminSubmissions:
      type: object
      required: [volunteers, materials, helpRequests, contacts]
      properties:
        volunteers:
          type: array
          items:
            type: object
            required: [id, name, phone, email, city, helpType, archived, createdAt]
            properties:
              id: { type: integer }
              name: { type: string }
              phone: { type: string }
              email: { type: string }
              city: { type: string }
              helpType: { type: string }
              archived: { type: boolean }
              createdAt: { type: string, format: date-time }
        materials:
          type: array
          items:
            type: object
            required: [id, name, phone, items, preferredDate, archived, createdAt]
            properties:
              id: { type: integer }
              name: { type: string }
              phone: { type: string }
              items: { type: string }
              preferredDate: { type: string }
              archived: { type: boolean }
              createdAt: { type: string, format: date-time }
        helpRequests:
          type: array
          items:
            type: object
            required: [id, childName, age, diagnosis, targetSum, parentContacts, archived, createdAt]
            properties:
              id: { type: integer }
              childName: { type: string }
              age: { type: integer }
              diagnosis: { type: string }
              targetSum: { type: integer }
              parentContacts: { type: string }
              photoUrl: { type: string, nullable: true }
              archived: { type: boolean }
              createdAt: { type: string, format: date-time }
        contacts:
          type: array
          items:
            type: object
            required: [id, name, email, message, archived, createdAt]
            properties:
              id: { type: integer }
              name: { type: string }
              email: { type: string }
              message: { type: string }
              archived: { type: boolean }
              createdAt: { type: string, format: date-time }

    DonationStats:
      type: object
      required: [perChild, totalWeek, totalAmount, recent]
      properties:
        totalWeek: { type: integer }
        totalAmount: { type: integer }
        perChild:
          type: array
          items:
            type: object
            required: [childId, childName, count, totalAmount]
            properties:
              childId: { type: integer }
              childName: { type: string }
              count: { type: integer }
              totalAmount: { type: integer }
        recent:
          type: array
          items:
            type: object
            required: [id, childId, childName, createdAt]
            properties:
              id: { type: integer }
              childId: { type: integer }
              childName: { type: string }
              amount: { type: integer, nullable: true }
              createdAt: { type: string, format: date-time }

    VisitStats:
      type: object
      required: [totalWeek, perDay, recent]
      properties:
        totalWeek: { type: integer }
        perDay:
          type: array
          items:
            type: object
            required: [date, count]
            properties:
              date: { type: string }
              count: { type: integer }
        recent:
          type: array
          items:
            type: object
            required: [id, createdAt]
            properties:
              id: { type: integer }
              path: { type: string }
              createdAt: { type: string, format: date-time }

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: "#/components/schemas/UploadUrlRequest" }